feat(admin, patient): add users page and patient side pages
Admin:
- Users page
  + Add user modal
  + User table

Patient:
- Added patient side pages

+++ database/migrations/2025_09_24_095221_alter_users_table.php
@@ -11,7 +11,9 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('name');
+            if (Schema::hasColumn('users', 'name')) {
+                $table->dropColumn('name');
+            }
         });
     }
 
@@ -21,7 +23,9 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('name')->nullable();
+            if (Schema::hasColumn('users', 'name')) {
+                $table->string('name')->nullable();
+            }
         });
     }
 };
